refactor(spec): remove import cycle and cleanup (#2831)
