+++ chain/helpers.go
@@ -56,3 +56,18 @@ func (s spec) SlotToEpoch(slot math.Slot) math.Epoch {
 func (s spec) WithinDAPeriod(block, current math.Slot) bool {
 	return s.SlotToEpoch(block)+s.MinEpochsForBlobsSidecarsRequest() >= s.SlotToEpoch(current)
 }
+
+// IsMainnet returns true if the chain is running with the mainnet chain ID.
+func (s spec) IsMainnet() bool {
+	return s.DepositEth1ChainID() == MainnetEth1ChainID
+}
+
+// IsTestnet returns true if the chain is running with the testnet chain ID.
+func (s spec) IsTestnet() bool {
+	return s.DepositEth1ChainID() == TestnetEth1ChainID
+}
+
+// IsDevnet returns true if the chain is running with the devnet chain ID.
+func (s spec) IsDevnet() bool {
+	return s.DepositEth1ChainID() == DevnetEth1ChainID
+}
